Added simple sbt build (also supporting scala.js).
Made it compatible with Scala 2.10-2.12.
Added comment about the sporadic test failures on JVM (though with that
change it still fails).
