+++ .gitignore
@@ -8,3 +8,6 @@ bin
 *~
 *.\#*
 *\#*\#
+.idea
+/project/project/target/
+/project/target/
