+++ AUTHORS
@@ -2,3 +2,4 @@ AUTHORS
 =======
 
 Michel Alexandre Salim <salimma@fedoraproject.org>
+Gabor Bakos
