+++ project/plugins.sbt
@@ -1,0 +1,3 @@
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.15")
+
+addSbtPlugin("org.tpolecat" % "tut-plugin" % "0.4.8")
