+++ shared/src/main/scala/info/hircus/kanren/Prelude.scala
