+++ shared/src/main/scala/info/hircus/kanren/examples/PalProd.scala
