+++ shared/src/main/scala/info/hircus/kanren/examples/SendMoreMoney.scala
