+++ shared/src/test/scala/info/hircus/kanren/tests/MathSpecification.scala
